+++ docs/changelog.rst
@@ -20,6 +20,9 @@ To update |kitty|, :doc:`follow the instructions <binary>`.
   could cause incorrect parsing if either the pending buffer capacity or the
   pending timeout were exceeded (:iss:`3779`)
 
+- Graphics protocol: Add support for composing rectangles from one animation
+  frame onto another (:iss:`3809`)
+
 - diff kitten: Remove limit on max line length of 4096 characters (:iss:`3806`)
 
 - Fix turning off cursor blink via escape codes not working (:iss:`3808`)